--- Teensy /CommboardV1.2/lib/mobrobI2C/mobrobI2C.cpp
@@ -1,2 +1,0 @@
-#include <i2c_t3.h>
-#include <Arduino.h>
